refactor: DTO setter 삭제
