+++ src/main/java/com/flab/gettoticket/dto/SeatDTO.java
@@ -1,8 +1,7 @@
 package com.flab.gettoticket.dto;
 
-import com.flab.gettoticket.model.PlayTime;
+import com.flab.gettoticket.entity.PlayTime;
 import lombok.Getter;
-import lombok.Setter;
 import lombok.ToString;
 
 import java.util.List;
